refactor(backoffice): isolate workflow UI interactions
Give generated workflow interfaces a focused interaction host contract.
Extract workflow-step generated UI and draft handling from the step
card,
and make hostless generated controls explicitly read-only.
